allow all students studying for a Computing degree to bear office
The current rule which requires candidates to be studying a course
with over 50% credit points in Computing units excludes students
of double degree courses which yield a Computing degree, such as
132210 and 132010.

In the context of double degree courses, units belonging to each
constituent degree are not specified.

A simple addition of an alternative criterion such as 'studying a
course administered by the School of Electrical Engineering and
Computing' is no longer possible because most of the popular
Computing majors have moved to being under the purview of the School
of Science. A course whitelist is thus unfortunately necessary.